Move RawRepresentable conformance to separate file
